Advection-Diffusion equation with Hughes' Diff-DG and SUPG
Created a new notebook from the inclusion of advection to the diffusion_Hughes notebook.
The formulation in this notebook is stable for a broad range of Peclet numbers. (Beware,
the formulation is not yet undimensional (needs to be done).
For low Peclet numbers, the diffusive terms have to be stabilized by choice of an appropriate
value for gamma ( gamma = 10 seemed to be a good choice for the zero rhs, for instance).
There is no need for this diffusion stabilization for high Peclet numbers (obviously), but it
won't hurt either.